+++ CHANGELOG.md
@@ -2,6 +2,77 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.8.4](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.8.3...v1.8.4) (2026-03-20)
+
+
+### Bug Fixes
+
+* upgrade Node.js runtime from 20 to 24 ([#457](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/457)) ([872f52a](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/872f52a952090f1e5c0a74571a9770401a664051)), closes [#454](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/454)
+
+### [1.8.3](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.8.2...v1.8.3) (2026-01-16)
+
+### [1.8.2](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.8.1...v1.8.2) (2025-12-18)
+
+### [1.8.1](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.8.0...v1.8.1) (2025-10-16)
+
+## [1.8.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.7.5...v1.8.0) (2025-08-14)
+
+
+### Features
+
+* add possibility to add more than one container in container-name ([#229](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/229)) ([ce1d739](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/ce1d7390795649b913f552685fa70aa7f8f90231))
+
+### [1.7.5](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.7.4...v1.7.5) (2025-08-04)
+
+### [1.7.4](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.7.3...v1.7.4) (2025-06-23)
+
+### [1.7.3](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.7.2...v1.7.3) (2025-05-28)
+
+### [1.7.2](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.7.1...v1.7.2) (2025-04-22)
+
+### [1.7.1](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.7.0...v1.7.1) (2025-03-24)
+
+## [1.7.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.6.2...v1.7.0) (2025-03-03)
+
+
+### Features
+
+* get tags and clean attributes ([#325](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/325)) ([ac0e284](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/ac0e284627e78d74f8161229324d67520fc20117))
+
+### [1.6.2](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.6.1...v1.6.2) (2025-01-14)
+
+### [1.6.1](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.6.0...v1.6.1) (2024-11-26)
+
+## [1.6.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.5.2...v1.6.0) (2024-10-23)
+
+
+### Features
+
+* support secrets in render task definition ([#334](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/334)) ([bd39926](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/bd399266e884fda2f863e083ff67d844c030f2b9))
+
+### [1.5.2](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.5.1...v1.5.2) (2024-10-01)
+
+### [1.5.1](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.5.0...v1.5.1) (2024-08-28)
+
+
+### Bug Fixes
+
+* Downgrade warning logs for fetch task definition feature ([#323](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/323)) ([b33b6ea](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/b33b6ea7210b1bbbd20b61f96a9c632d4886465a))
+
+## [1.5.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.4.0...v1.5.0) (2024-08-08)
+
+
+### Features
+
+* add new feature to fetch task definition using task definition arn or name([#311](https://github.com/aws-actions/amazon-ecs-render-task-definition/issues/311)) ([e246ecd](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/e246ecdd89c8de5a4a8c248830566528cbed06dd))
+
+## [1.4.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.3.0...v1.4.0) (2024-07-18)
+
+
+### Features
+
+* Adds env-files input to the github action ([a40f6ea](https://github.com/aws-actions/amazon-ecs-render-task-definition/commit/a40f6eaad3be711b605d54fefc660fbd40925fdb))
+
 ## [1.3.0](https://github.com/aws-actions/amazon-ecs-render-task-definition/compare/v1.2.0...v1.3.0) (2024-05-07)